refactor: client folder structure
