+++ client/src/components/layout/BlogItem.tsx
