+++ client/src/components/layout/profile/ProfileAbout.tsx
@@ -1,4 +1,4 @@
-import stitches from '../../stitches.config';
+import stitches from '../../../stitches.config';
 
 const { styled } = stitches;
