+++ client/src/pages/user/Profile.tsx
@@ -2,9 +2,9 @@ import stitches from '../../stitches.config';
 import NotFound from '../NotFound';
 import { IUser } from '../../types/user';
 import { IUserIcon } from '../../types/userIcon';
-import { A as NavLink } from '../../components/stitches/elements';
-import ProfileAbout from '../../components/layout/ProfileAbout';
-import BlogItem from '../../components/blog/BlogItem';
+import { A as NavLink } from '../../components/common/elements';
+import ProfileAbout from '../../components/layout/profile/ProfileAbout';
+import BlogItem from '../../components/layout/BlogItem';
 
 import { useEffect, useState } from 'react';
 import { useParams, Link, Route, Routes, useMatch } from 'react-router-dom';
